seranking/requirements.txt
requests==2.31.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1
python-dotenv==1.0.0
jupyter==1.0.0
ipykernel==6.28.0
openpyxl==3.1.2
